# public/swagger-spec/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"FileUploadSuccessResponseReasonEnum": {
				"enum": [
					"Upload was successful"
				],
				"type": "string"
			},
			"StatusCodes": {
				"enum": [
					100,
					101,
					102,
					103,
					200,
					201,
					202,
					203,
					204,
					205,
					206,
					207,
					300,
					301,
					302,
					303,
					304,
					305,
					307,
					308,
					400,
					401,
					402,
					403,
					404,
					405,
					406,
					407,
					408,
					409,
					410,
					411,
					412,
					413,
					414,
					415,
					416,
					417,
					418,
					419,
					420,
					421,
					422,
					423,
					424,
					426,
					428,
					429,
					431,
					451,
					500,
					501,
					502,
					503,
					504,
					505,
					507,
					511
				],
				"type": "number"
			},
			"FileUploadSuccessResponse": {
				"properties": {
					"message": {
						"$ref": "#/components/schemas/FileUploadSuccessResponseReasonEnum"
					},
					"status": {
						"$ref": "#/components/schemas/StatusCodes"
					}
				},
				"required": [
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasicAuthFailureResponseReasonEnum": {
				"enum": [
					"Authorization header contains invalid credentials",
					"Authorization header is malformed",
					"Authorization header is missing"
				],
				"type": "string"
			},
			"FileUploadFailureResponseReasonEnum": {
				"enum": [
					"Could not upload all file(s)",
					"Could not upload file(s)",
					"Exceeded the allowed file size. Try again with a smaller file",
					"Exceeded the allowed number of file(s) that can be uploaded in a single request",
					"Exceeded the allowed number of requests. Try again in a few minutes.",
					"Invalid file",
					"No file(s) to upload"
				],
				"type": "string"
			},
			"FileUploadFailureResponse": {
				"properties": {
					"reason": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/BasicAuthFailureResponseReasonEnum"
							},
							{
								"$ref": "#/components/schemas/FileUploadFailureResponseReasonEnum"
							}
						]
					},
					"status": {
						"$ref": "#/components/schemas/StatusCodes"
					}
				},
				"required": [
					"reason",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"basic_auth": {
				"in": "headers",
				"scheme": "basic",
				"type": "http"
			}
		}
	},
	"info": {
		"title": "file-uploader-rest-service",
		"version": "1.0.0",
		"description": "The backend REST service that handles requests to upload files to filesystem",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/file-upload": {
			"post": {
				"operationId": "file-upload",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/FileUploadSuccessResponse"
										},
										{
											"$ref": "#/components/schemas/FileUploadFailureResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "No file(s) to upload",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FileUploadFailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "Authorization header is missing",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FileUploadFailureResponse"
								}
							}
						}
					},
					"429": {
						"description": "Exceeded the allowed number of requests. Try again in a few minutes.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FileUploadFailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "Could not upload file(s)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FileUploadFailureResponse"
								}
							}
						}
					}
				},
				"description": "Endpoint responsible for handling file uploads requests from client.\nIt will return a message or a reason indicating why it has failed.\nDon't forget to add the authorization header to the request with adequate base64 encoded credentials\nFor example, if service's credentials are username: 'user' and password: '123456', then you should send the following: Authorization: Basic dXNlcjoxMjM0NTY=\nYou can use https://www.base64encode.org/ to quickly encode your credentials and https://www.base64decode.org/ to quickly decode them\nIf you don't have such credentials, you should request them from the team.",
				"tags": [
					"Handle file uploads requests from client"
				],
				"security": [
					{
						"basic_auth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"files": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"files"
								]
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}